Parallel version

+++ project/code/param_config.h
@@ -4,23 +4,7 @@
 #include "zf_common_typedef.h"
 
 #include "PID.h"
-// pid结构体，位于"PID.h"（只是体现结构，不要解除注释）
-//typedef struct {
-//	float Target;
-//	float Actual;
-//	float Out;
-//	
-//	float Kp;
-//	float Ki;
-//	float Kd;
-//	
-//	float Error0;
-//	float Error1;
-//	float ErrorInt;
-//	
-//	float OutMax;
-//	float OutMin;
-//} PID_t;
+// 注：pid结构体，位于"PID.h"
 
 //定义PID
 extern PID_t Angle_PID;					//角度环
@@ -43,7 +27,7 @@ extern int16_t LeftPWM;					//左轮PWM
 extern int16_t RightPWM;				//右轮PWM
 extern int16_t AvePWM;					//平均PWM
 extern int16_t DifPWM;					//差分PWM
-extern float LeftSpeed, RightSpeed;		//左轮实际速度，右轮实际速度
+extern int16_t LeftSpeed, RightSpeed;		//左轮实际速度，右轮实际速度
 extern float AveSpeed, DifSpeed;		//平均实际速度，差分实际速度
 
 // 调控周期计时器
